+++ packages/expo-modules-autolinking/src/dependencies/__tests__/resolution-test.ts
@@ -57,6 +57,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-third-party",
           "originPath": "/fake/project/node_modules/react-native-third-party",
           "path": "/fake/project/node_modules/react-native-third-party",
+          "source": 0,
           "version": "0.0.1",
         },
       }
@@ -95,6 +96,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-dependency",
           "originPath": "/fake/project/node_modules/react-native-third-party/node_modules/react-native-dependency",
           "path": "/fake/project/node_modules/react-native-third-party/node_modules/react-native-dependency",
+          "source": 0,
           "version": "0.0.1",
         },
         "react-native-third-party": {
@@ -103,6 +105,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-third-party",
           "originPath": "/fake/project/node_modules/react-native-third-party",
           "path": "/fake/project/node_modules/react-native-third-party",
+          "source": 0,
           "version": "0.0.1",
         },
       }
@@ -135,6 +138,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-dependency",
           "originPath": "/fake/project/node_modules/react-native-dependency",
           "path": "/fake/project/node_modules/react-native-dependency",
+          "source": 0,
           "version": "0.0.1",
         },
         "react-native-third-party": {
@@ -143,6 +147,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-third-party",
           "originPath": "/fake/project/node_modules/react-native-third-party",
           "path": "/fake/project/node_modules/react-native-third-party",
+          "source": 0,
           "version": "0.0.1",
         },
       }
@@ -188,6 +193,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-dependency",
           "originPath": "/fake/project/node_modules/.pnpm/react-native-third-party@x.x.x/node_modules/react-native-dependency",
           "path": "/fake/project/node_modules/.pnpm/react-native-dependency@x.x.x/node_modules/react-native-dependency",
+          "source": 0,
           "version": "0.0.1",
         },
         "react-native-third-party": {
@@ -196,6 +202,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-third-party",
           "originPath": "/fake/project/node_modules/react-native-third-party",
           "path": "/fake/project/node_modules/.pnpm/react-native-third-party@x.x.x/node_modules/react-native-third-party",
+          "source": 0,
           "version": "0.0.1",
         },
       }
@@ -226,6 +233,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-third-party",
           "originPath": "/fake/project/node_modules/react-native-third-party",
           "path": "/fake/project/node_modules/react-native-third-party",
+          "source": 0,
           "version": "0.0.1",
         },
       }
@@ -258,6 +266,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "included",
           "originPath": "/fake/project/node_modules/included",
           "path": "/fake/project/node_modules/included",
+          "source": 0,
           "version": "0.0.1",
         },
       }
@@ -302,6 +311,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "parent-a",
           "originPath": "/fake/project/node_modules/parent-a",
           "path": "/fake/project/node_modules/parent-a",
+          "source": 0,
           "version": "0.0.1",
         },
         "parent-b": {
@@ -310,16 +320,23 @@ describe(scanDependenciesRecursively, () => {
           "name": "parent-b",
           "originPath": "/fake/project/node_modules/parent-b",
           "path": "/fake/project/node_modules/parent-b",
+          "source": 0,
           "version": "0.0.1",
         },
         "react-native-dependency": {
           "depth": 1,
           "duplicates": [
-            "/fake/project/node_modules/parent-b/node_modules/react-native-dependency",
+            {
+              "name": "react-native-dependency",
+              "originPath": "/fake/project/node_modules/parent-b/node_modules/react-native-dependency",
+              "path": "/fake/project/node_modules/parent-b/node_modules/react-native-dependency",
+              "version": "",
+            },
           ],
           "name": "react-native-dependency",
           "originPath": "/fake/project/node_modules/parent-a/node_modules/react-native-dependency",
           "path": "/fake/project/node_modules/parent-a/node_modules/react-native-dependency",
+          "source": 0,
           "version": "0.0.1",
         },
       }
@@ -358,6 +375,7 @@ describe(scanDependenciesRecursively, () => {
           "name": "react-native-third-party",
           "originPath": "/fake/project/node_modules/react-native-third-party",
           "path": "/fake/project/node_modules/react-native-third-party",
+          "source": 0,
           "version": "",
         },
       }